Create README.md
This was a fun challenge to complete from @frontendmentor. This is a clean and simple responsive/interactive rating component that can be added seamlessly to any web page. I may or may not edit this code in the future with more stuff, and this code may also break in the future, who knows.